refactor: update file paths for button components
