+++ src/components/buttons/index.ts
@@ -1,0 +1,5 @@
+import IconButton from './IconButton';
+import SubmitButton from './SubmitButton';
+import TextButton from './TextButton';
+
+export { IconButton, SubmitButton, TextButton };
